Moved LayoutRes and StringRes classes into "view" package
