+++ client/src/main/scala/it/cwmp/client/view/LayoutRes.scala
@@ -1,4 +1,4 @@
-package it.cwmp.client.utils
+package it.cwmp.client.view
 
 /**
   * Utility class that keeps together all the file names of the FXML layout files.
