+++ client/src/main/scala/it/cwmp/client/view/game/GameFX.scala
@@ -2,9 +2,8 @@ package it.cwmp.client.view.game
 
 import akka.actor.ActorRef
 import it.cwmp.client.model.game.impl.{Cell, CellWorld, Point}
-import it.cwmp.client.utils.LayoutRes
 import it.cwmp.client.view.game.model.{CellView, TentacleView}
-import it.cwmp.client.view.{FXAlertsController, FXViewController}
+import it.cwmp.client.view.{FXAlertsController, FXViewController, LayoutRes}
 import javafx.fxml.FXML
 import javafx.scene.canvas.Canvas
 import javafx.scene.input.MouseEvent
